+++ packages/cli/src/index.ts
@@ -3,5 +3,5 @@ export type { ConfigOptions } from "@lage-run/config";
 export type { Priority } from "@lage-run/config";
 export type { PipelineDefinition } from "@lage-run/config";
 export type { LoggerOptions } from "@lage-run/config";
-export type { TargetRunnerPickerOptions } from "@lage-run/scheduler-types";
+export type { TargetRunnerPickerOptions } from "@lage-run/runners";
 export type { TargetConfig } from "@lage-run/target-graph";
