+++ packages/config/package.json
@@ -21,6 +21,7 @@
   },
   "dependencies": {
     "@lage-run/logger": "^1.3.0",
+    "@lage-run/runners": "^1.0.0",
     "@lage-run/scheduler-types": "^0.3.13",
     "@lage-run/target-graph": "^0.8.9",
     "backfill-config": "6.4.2",
