--- packages/runners/src/types/TargetStatus.ts
@@ -1,1 +1,0 @@
-export type TargetStatus = "pending" | "queued" | "running" | "success" | "failed" | "skipped" | "aborted";
