--- packages/scheduler/src/runners/NoOpRunner.ts
@@ -1,11 +1,0 @@
-import type { TargetRunner } from "@lage-run/scheduler-types";
-
-export class NoOpRunner implements TargetRunner {
-  async shouldRun() {
-    return true;
-  }
-
-  async run() {
-    // pass
-  }
-}
